Added switch methods, turn end check methods, and moved some methods or another file
